Added a testing program for QueryIndex.
This program is housed within its own directory in the insights
folder and is able to automatically run test cases that verify
the functional correctness of QueryIndex's implementation, as
written in tensor2tensor/insights/indexing.py.

The tests depend on the presence of certain data files which are
too big to add to this repository. To obtain these files, run
wget www.statmt.org/europarl/v7/de-en.tgz and extract the files
into the index_testing directory.

Currently, the test cases only verify the searching functionality
and are limited to a single data test case. This is because changes
to the interface of QueryIndex are expected as of this update;
as soon as those changes are made, tests for index creation and
retrieval and well as testing of the class interface will be added.